app/requirements.txt
pyserial
pygame
colorama
matplotlib
numpy
PyQt5
tk